Moves p2p config to file
